Reduce usage of tabes
