+++ pyqrcode/tables.py
@@ -64,16 +64,6 @@ ascii_codes = {'0': 0, '1': 1, '2': 2, '3': 3, '4': 4, '5': 5, '6': 6, '7': 7,
                ' ': 36, '$': 37, '%': 38, '*': 39, '+': 40, '-': 41, '.': 42,
                '/': 43, ':': 44}
 
-#: This array specifies the size of a QR Code in pixels. These numbers are
-#: defined in the standard. The indexes correspond to the QR Code's
-#: version number. This array was taken from:
-#:
-#: http://www.denso-wave.com/qrcode/vertable1-e.html
-version_size = [None, 21, 25, 29, 33, 37, 41, 45, 49, 53, 57,
-                61, 65, 69, 73, 77, 81, 85, 89, 93, 97,
-                101, 105, 109, 113, 117, 121, 125, 129, 133, 137,
-                141, 145, 149, 153, 157, 161, 165, 169, 173, 177]
-
 #: This dictionary lists the data capacity for all possible QR Codes.
 #: This dictionary is organized where the first key corresponds to the
 #: QR Code version number. The next key corresponds to the error
